requirements.txt
openai==1.3.0
instructor==0.3.4
deepdiff==6.7.1
termcolor==2.3.0
python-dotenv==1.0.0
pydantic
gradio==3.41.0